cast floats to ints when calling resize(), setX(), and setY() (argument types changed in newer version)

+++ src/gui/preview_win.py
@@ -25,8 +25,8 @@ class PreviewWindow(QtWidgets.QLabel):
             transformMode=QtCore.Qt.SmoothTransformation)
 
         # start painting the label from left upper corner
-        point.setX((size.width() - scaledPix.width())/2)
-        point.setY((size.height() - scaledPix.height())/2)
+        point.setX(int((size.width() - scaledPix.width())/2))
+        point.setY(int((size.height() - scaledPix.height())/2))
         painter.drawPixmap(point, scaledPix)
 
     def changePixmap(self, img):
